NMS-9128: Move tests to same packages as classes the tests are for
