+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.asset/src/test/java/org/opennms/features/topology/plugins/topo/asset/util/NodeBuilder.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.features.topology.plugins.topo.asset.test;
+package org.opennms.features.topology.plugins.topo.asset.util;
 
 import java.util.Arrays;
 import java.util.Set;
